Refactor into smaller components
refs #49
